Added new items. Added new map, Shamrock Harbour. Tweaked enemy stats on Azureal Island. Changed Entity positions to doubles. Added new sound effects for impacts.

+++ src/dev/ipsych0/myrinnia/abilities/PBAoECircle.java
@@ -7,6 +7,6 @@ import java.awt.geom.Ellipse2D;
 class PBAoECircle extends Ellipse2D.Float {
 
     public PBAoECircle(int circleSize) {
-        super(Handler.get().getPlayer().getX() + Handler.get().getPlayer().getWidth() / 2f - circleSize / 2f, Handler.get().getPlayer().getY() + Handler.get().getPlayer().getHeight() / 2f - circleSize / 2f, circleSize, circleSize);
+        super((float) Handler.get().getPlayer().getX() + Handler.get().getPlayer().getWidth() / 2f - circleSize / 2f, (float) Handler.get().getPlayer().getY() + Handler.get().getPlayer().getHeight() / 2f - circleSize / 2f, circleSize, circleSize);
     }
 }
