+++ src/dev/ipsych0/myrinnia/entities/statics/StaticEntity.java
@@ -13,7 +13,7 @@ public abstract class StaticEntity extends Entity {
      */
     private static final long serialVersionUID = -2206779374852046145L;
 
-    protected StaticEntity(float x, float y, int width, int height, String name, int level, String dropTable, String jsonFile, String animation, String itemsShop) {
+    protected StaticEntity(double x, double y, int width, int height, String name, int level, String dropTable, String jsonFile, String animation, String itemsShop) {
         super(x, y, width, height, name, level, dropTable, jsonFile, animation, itemsShop);
 
         staticNpc = true;
