+++ src/dev/ipsych0/myrinnia/gfx/GameCamera.java
@@ -13,9 +13,9 @@ public class GameCamera implements Serializable {
      *
      */
     private static final long serialVersionUID = 6890827040135434870L;
-    private float xOffset, yOffset;
+    private double xOffset, yOffset;
 
-    public GameCamera(float xOffset, float yOffset) {
+    public GameCamera(double xOffset, double yOffset) {
         this.xOffset = xOffset;
         this.yOffset = yOffset;
     }
@@ -40,25 +40,25 @@ public class GameCamera implements Serializable {
         checkBlankSpace();
     }
 
-    public void move(float xAmount, float yAmount) {
+    public void move(double xAmount, double yAmount) {
         xOffset += xAmount;
         yOffset += yAmount;
         checkBlankSpace();
     }
 
-    public float getxOffset() {
+    public double getxOffset() {
         return xOffset;
     }
 
-    public void setxOffset(float xOffset) {
+    public void setxOffset(double xOffset) {
         this.xOffset = xOffset;
     }
 
-    public float getyOffset() {
+    public double getyOffset() {
         return yOffset;
     }
 
-    public void setyOffset(float yOffset) {
+    public void setyOffset(double yOffset) {
         this.yOffset = yOffset;
     }
